Update head.html
When you add Watcher3 to the homescreen on iOS, it basically just shows a small screenshot of the webpage as the icon, which doesn't look very good in icon form, and also changes all the time. Adding this line will tell iOS where to look to get the proper icon. You will also need to add the icon file which I uploaded to https://filetea.me/n3wbvMxmy6qSHmHIlDxtTkWNQ, and copy it to /static/images/apple-touch-icon.png